mv gene_vcf
